backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
google-cloud-documentai==2.20.1
google-auth==2.23.4
pydantic==1.10.13
pydantic-settings==2.1.0
python-multipart==0.0.6

